# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Information Crawler API",
    "summary": "中关村人工智能研究院 — 信息监测系统",
    "description": "## 概述\n\n信息监测系统 API，自动爬取 **129 个信源**（105 个启用），覆盖 **9 个维度**，为中关村人工智能研究院提供全方位的信息监测与商业智能服务。\n\n## 功能模块\n\n| 模块 | 说明 |\n|------|------|\n| **文章管理** | 全量文章的查询、搜索、统计 |\n| **信源管理** | 129 个信源的配置、状态监控、手动触发 |\n| **维度视图** | 9 大维度的文章聚合浏览 |\n| **政策智能** | 规则引擎 + LLM 二级管线，政策机会挖掘 |\n| **人事情报** | 任免信息自动提取，LLM 相关性分析 |\n| **系统健康** | 数据库、调度器、爬取健康度监控 |\n\n## 维度说明\n\n- `national_policy` — 国家政策（国务院、部委）\n- `beijing_policy` — 北京政策（市/区政府）\n- `technology` — 技术动态（ArXiv、GitHub Trending、Hacker News 等）\n- `talent` — 人才政策\n- `industry` — 产业动态\n- `universities` — 高校动态（46 所高校 AI 院系）\n- `events` — 活动会议\n- `personnel` — 人事变动\n- `twitter` — Twitter/X KOL 动态\n\n## 技术栈\n\nFastAPI + SQLAlchemy(async) + PostgreSQL(Supabase) + APScheduler + httpx + BeautifulSoup4 + Playwright",
    "contact": {
      "name": "中关村人工智能研究院",
      "url": "https://www.zgcaiia.com/"
    },
    "license": {
      "name": "Internal Use"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sources/": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "信源列表",
        "description": "查询所有信源及其状态信息，可按维度过滤。",
        "operationId": "list_sources_api_v1_sources__get",
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dimension"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceResponse"
                  },
                  "title": "Response List Sources Api V1 Sources  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "信源详情",
        "description": "根据信源 ID 获取详细配置和运行状态。",
        "operationId": "get_source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "404": {
            "description": "信源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sources"
        ],
        "summary": "启用/禁用信源",
        "description": "更新信源的启用状态。禁用后调度器将不再为该信源创建爬取任务。",
        "operationId": "update_source_api_v1_sources__source_id__patch",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "404": {
            "description": "信源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/logs": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "爬取日志",
        "description": "获取指定信源的最近爬取日志，默认返回最近 20 条。",
        "operationId": "get_source_logs_api_v1_sources__source_id__logs_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlLogResponse"
                  },
                  "title": "Response Get Source Logs Api V1 Sources  Source Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/trigger": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "手动触发爬取",
        "description": "立即触发指定信源的一次爬取任务，不影响定时调度。",
        "operationId": "trigger_crawl_api_v1_sources__source_id__trigger_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "信源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "调度器未运行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/intel/policy/feed": {
      "get": {
        "tags": [
          "intel",
          "policy-intel"
        ],
        "summary": "政策动态 Feed",
        "description": "获取政策情报动态列表。支持按分类、重要性、匹配度过滤，以及关键词搜索。\n\n数据经过规则引擎 + LLM 二级管线处理，包含匹配度评分、资金信息、AI 洞察等富化字段。",
        "operationId": "get_policy_feed_api_v1_intel_policy_feed_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "政策分类: 国家政策 / 北京政策 / 领导讲话 / 政策机会",
              "title": "Category"
            },
            "description": "政策分类: 国家政策 / 北京政策 / 领导讲话 / 政策机会"
          },
          {
            "name": "importance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "重要性级别: 紧急 / 重要 / 关注 / 一般",
              "title": "Importance"
            },
            "description": "重要性级别: 紧急 / 重要 / 关注 / 一般"
          },
          {
            "name": "min_match_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最低匹配度得分",
              "title": "Min Match Score"
            },
            "description": "最低匹配度得分"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键词搜索（标题/摘要/来源/标签）",
              "title": "Keyword"
            },
            "description": "关键词搜索（标题/摘要/来源/标签）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "返回条数上限",
              "default": 50,
              "title": "Limit"
            },
            "description": "返回条数上限"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intel/policy/opportunities": {
      "get": {
        "tags": [
          "intel",
          "policy-intel"
        ],
        "summary": "政策机会看板",
        "description": "获取政策申报机会列表，用于情报看板展示。包含资金规模、截止日期、匹配度评分、紧急状态等关键字段。",
        "operationId": "get_policy_opportunities_api_v1_intel_policy_opportunities_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态过滤: urgent / active / tracking",
              "title": "Status"
            },
            "description": "状态过滤: urgent / active / tracking"
          },
          {
            "name": "min_match_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最低匹配度得分",
              "title": "Min Match Score"
            },
            "description": "最低匹配度得分"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "返回条数上限",
              "default": 50,
              "title": "Limit"
            },
            "description": "返回条数上限"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOpportunitiesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrawlLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "日志 ID",
            "examples": [
              5678
            ]
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "description": "信源 ID",
            "examples": [
              "tech_arxiv"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "爬取状态: success / partial / failed",
            "examples": [
              "success"
            ]
          },
          "items_total": {
            "type": "integer",
            "title": "Items Total",
            "description": "本次解析到的条目总数",
            "examples": [
              25
            ]
          },
          "items_new": {
            "type": "integer",
            "title": "Items New",
            "description": "新增入库的条目数",
            "examples": [
              12
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "错误信息（成功时为 null）"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "任务开始时间"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At",
            "description": "任务结束时间"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "耗时（秒）",
            "examples": [
              3.45
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "status",
          "items_total",
          "items_new",
          "started_at"
        ],
        "title": "CrawlLogResponse",
        "description": "单次爬取任务的日志记录。"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "错误详情",
            "examples": [
              "Article not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyFeedItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "记录 ID",
            "examples": [
              "pol_20240115_001"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "政策标题",
            "examples": [
              "关于印发《新一代人工智能发展规划》的通知"
            ]
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "政策摘要"
          },
          "category": {
            "type": "string",
            "enum": [
              "国家政策",
              "北京政策",
              "领导讲话",
              "政策机会"
            ],
            "title": "Category",
            "description": "政策分类"
          },
          "importance": {
            "type": "string",
            "enum": [
              "紧急",
              "重要",
              "关注",
              "一般"
            ],
            "title": "Importance",
            "description": "重要性级别"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "description": "发布日期",
            "examples": [
              "2024-01-15"
            ]
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "来源名称",
            "examples": [
              "国务院"
            ]
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "标签",
            "default": [],
            "examples": [
              [
                "人工智能",
                "规划"
              ]
            ]
          },
          "matchScore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matchscore",
            "description": "与研究院业务的匹配度（0-100）",
            "examples": [
              92
            ]
          },
          "funding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding",
            "description": "涉及资金规模",
            "examples": [
              "5000万元"
            ]
          },
          "daysLeft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daysleft",
            "description": "截止日倒计时（天）",
            "examples": [
              15
            ]
          },
          "leader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leader",
            "description": "相关领导",
            "examples": [
              "张某某"
            ]
          },
          "relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance",
            "description": "LLM 评估的相关性得分（0-100）",
            "examples": [
              88
            ]
          },
          "signals": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signals",
            "description": "关键信号词列表",
            "examples": [
              [
                "人工智能",
                "专项资金",
                "中关村"
              ]
            ]
          },
          "sourceUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceurl",
            "description": "原文链接"
          },
          "aiInsight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aiinsight",
            "description": "AI 分析洞察"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "详细说明"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "原文正文"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "summary",
          "category",
          "importance",
          "date",
          "source"
        ],
        "title": "PolicyFeedItem",
        "description": "政策动态 Feed 中的单条记录。"
      },
      "PolicyFeedResponse": {
        "properties": {
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "数据生成时间",
            "examples": [
              "2024-01-15T10:30:00"
            ]
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count",
            "description": "返回的记录数",
            "examples": [
              25
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PolicyFeedItem"
            },
            "type": "array",
            "title": "Items",
            "description": "政策动态列表"
          }
        },
        "type": "object",
        "required": [
          "item_count",
          "items"
        ],
        "title": "PolicyFeedResponse",
        "description": "政策动态 Feed 响应。"
      },
...
      "PolicyOpportunitiesResponse": {
        "properties": {
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "数据生成时间",
            "examples": [
              "2024-01-15T10:30:00"
            ]
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count",
            "description": "返回的记录数",
            "examples": [
              8
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PolicyItem"
            },
            "type": "array",
            "title": "Items",
            "description": "政策机会列表"
          }
        },
        "type": "object",
        "required": [
          "item_count",
          "items"
        ],
        "title": "PolicyOpportunitiesResponse",
        "description": "政策机会列表响应。"
      },
      "SourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "信源唯一 ID",
            "examples": [
              "tech_arxiv"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "信源名称",
            "examples": [
              "ArXiv AI Papers"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "信源 URL",
            "examples": [
              "https://arxiv.org/list/cs.AI/recent"
            ]
          },
          "dimension": {
            "type": "string",
            "title": "Dimension",
            "description": "所属维度",
            "examples": [
              "technology"
            ]
          },
          "crawl_method": {
            "type": "string",
            "title": "Crawl Method",
            "description": "爬取方式: static / dynamic / rss / snapshot / custom",
            "examples": [
              "static"
            ]
          },
          "schedule": {
            "type": "string",
            "title": "Schedule",
            "description": "调度频率: hourly / daily / twice_daily 等",
            "examples": [
              "daily"
            ]
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "description": "是否启用",
            "examples": [
              true
            ]
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "优先级（1-5，1 最高）",
            "examples": [
              2
            ]
          },
          "last_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Crawl At",
            "description": "上次爬取时间"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At",
            "description": "上次成功爬取时间"
          },
          "consecutive_failures": {
            "type": "integer",
            "title": "Consecutive Failures",
            "description": "连续失败次数",
            "default": 0,
            "examples": [
              0
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "url",
          "dimension",
          "crawl_method",
          "schedule",
          "is_enabled",
          "priority"
        ],
        "title": "SourceResponse",
        "description": "信源详情。"
      },
      "SourceUpdate": {
        "properties": {
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled",
            "description": "启用或禁用信源"
          }
        },
        "type": "object",
        "title": "SourceUpdate",
        "description": "信源更新请求体。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}